Adding the raw_url, so the next step is to download the most changed files and calculate their cognitive complexity

+++ src/report.rs
@@ -1,13 +1,10 @@
-
-
-pub fn print_report_without_header(top_changed_files: &Vec<(String, usize)>) {
-    top_changed_files.iter()
-    .for_each(|(file,num_changes)|{
-        println!("{}\t{}",file, num_changes);
+pub fn print_report_without_header(top_changed_files: &Vec<(String, u32)>) {
+    top_changed_files.iter().for_each(|(file, num_changes)| {
+        println!("{}\t{}", file, num_changes);
     });
 }
 
-pub fn print_report(top_changed_files: &Vec<(String, usize)>) {
+pub fn print_report(top_changed_files: &Vec<(String, u32)>) {
     println!("{}", format!("{:80}", "-").replace(" ", "-"));
     println!("File\t\tNumber of changes");
     println!("{}", format!("{:80}", "-").replace(" ", "-"));
